IOS-5864 refactor: extract shared WidgetChatPreviewBuilder
