+++ Anytype/Sources/PresentationLayer/Modules/HomeWidgets/Widgets/Builders/WidgetRowModelBuilder.swift
@@ -18,8 +18,8 @@ protocol WidgetRowModelBuilderProtocol: AnyObject, Sendable {
 @MainActor
 final class WidgetRowModelBuilder: WidgetRowModelBuilderProtocol, Sendable {
 
-    private let dateFormatter = ChatPreviewDateFormatter()
-    
+    private let chatPreviewBuilder: any WidgetChatPreviewBuilderProtocol = Container.shared.widgetChatPreviewBuilder()
+
     func buildGalleryRows(
         from configs: [SetContentViewItemConfiguration]
     ) -> [GalleryWidgetRowModel] {
@@ -32,46 +32,13 @@ final class WidgetRowModelBuilder: WidgetRowModelBuilderProtocol, Sendable {
         chatPreviews: [ChatMessagePreview]
     ) -> [ListWidgetRowModel] {
         configs.map { config in
-            let chatPreview = buildChatPreview(
+            let chatPreview = chatPreviewBuilder.build(
+                chatPreviews: chatPreviews,
                 objectId: config.id,
-                spaceView: spaceView,
-                chatPreviews: chatPreviews
+                spaceView: spaceView
             )
             return ListWidgetRowModel(details: config, chatPreview: chatPreview)
         }
-    }
-
-    private func buildChatPreview(
-        objectId: String,
-        spaceView: SpaceView?,
-        chatPreviews: [ChatMessagePreview]
-    ) -> MessagePreviewModel? {
-        guard let preview = chatPreviews.first(where: { $0.chatId == objectId }),
-              let lastMessage = preview.lastMessage else {
-            return nil
-        }
-
-        let attachments = lastMessage.attachments.prefix(3).map { objectDetails in
-            MessagePreviewModel.Attachment(
-                id: objectDetails.id,
-                icon: objectDetails.objectIconImage
-            )
-        }
-
-        let notificationMode = spaceView?.effectiveNotificationMode(for: objectId) ?? .all
-
-        return MessagePreviewModel(
-            creatorTitle: lastMessage.creator?.title,
-            text: lastMessage.text,
-            attachments: Array(attachments),
-            localizedAttachmentsText: lastMessage.localizedAttachmentsText,
-            chatPreviewDate: dateFormatter.localizedDateString(for: lastMessage.createdAt, showTodayTime: true),
-            unreadCounter: preview.unreadCounter,
-            mentionCounter: preview.mentionCounter,
-            hasUnreadReactions: preview.hasUnreadReactions,
-            notificationMode: notificationMode,
-            chatName: nil
-        )
     }
 }
